doc: improved doc
Add the date the project was started on the local machine in addition to the github repo creation date. Add a more descriptive title and additionally list numpy documentation in the credits.

The title and dates are more specific which help in getting a better idea of the project. Including the numpy documentation also lists the libraries used