refactor: clean get_filters()
rewrite the conditional for city input in get_filters(). Use the same condition style in city input as the rest using a new city_list variables. Declare city_list, week_day_list and month_list as global variables at the start of the code.

This changes make get filters easier to read and makes it easier to add or remove future elements from the user input lists by declaring them at the start of the script and not inside a function